LayoutTests/fast/forms/select-multiple-elements-with-mouse-drag-expected.txt

Select multiple option with mouse drag


PASS document.getElementById("selectId").options[0].selected is true
PASS document.getElementById("selectId").options[1].selected is true
PASS document.getElementById("selectId").options[2].selected is true
PASS document.getElementById("selectId").options[3].selected is true
PASS successfullyParsed is true

TEST COMPLETE
